vasads

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -15,7 +15,7 @@ int sum_them_all(const unsigned int n, ...)
 	va_start(args, n);
 	while ((unsigned int)i < n)
 	{
-		sum += va_arg(args, int);
+		(unsigned int)sum += va_arg(args, int);
 		i++;
 	}
 	va_end(args);
